Add components, React structure.
