+++ client/src/App.css
@@ -2,14 +2,6 @@
   text-align: center;
 }
 
-.App-logo {
-  animation: App-logo-spin infinite 20s linear;
-  height: 40vmin;
-  pointer-events: none;
-}
-
-
-
 .App-header {
   background-color: white;
   min-height: 100vh;
@@ -19,22 +11,4 @@
   justify-content: center;
   font-size: calc(10px + 2vmin);
 
-}
-
-
-.App-link {
-  color: #61dafb;
-}
-/* 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-} */
-
-p {
-  font-size: 2em;
 }
