--- client/src/components/AgeGate.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const AgeGate = () => {
-  return (
-    <p>This is our age gate</p>
-  )
-}
-
-export default AgeGate;
