--- client/src/components/Home.jsx
@@ -1,51 +1,0 @@
-import React from 'react';
-import { Route, Link, withRouter } from 'react-router-dom';
-import RegisterForm from './RegisterForm';
-import LoginForm from './LoginForm';
-import User from './User';
-
-class Home extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      register: false,
-      login: false
-    }
-  }
-  handleRegister = () => {
-    this.setState({
-      register: true,
-      login: false
-    })
-  }
-  handleLogin = () => {
-    this.setState({
-      login: true,
-      register: false
-    })
-  }
-
-  goToUser = () => {
-    this.props.history.push('/user')
-  }
-  render() {
-    return (
-      <div>
-        <nav>
-          <Link to='/home/register' onClick={this.handleRegister}>Register</Link>
-          <Link to='/home/login' onClick={this.handleLogin}>Log In</Link>
-        </nav>
-        <main>
-          <p>This is our homepage</p>
-          {this.state.register &&
-            <Route path='/home/register' render={() => <RegisterForm handleSubmit={this.goToUser} />} />}
-          {this.state.login &&
-            <Route path='/home/login' render={() => <LoginForm handleSubmit={this.goToUser} />} />}
-        </main>
-
-
-      </div>
-    )
-  }
-}
-export default withRouter(Home)
