--- client/src/components/User.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { Route } from 'react-router-dom';
-
-class User extends React.Component {
-  render() {
-    return (
-      <main>
-        <p>This is the user page</p>
-      </main>
-    )
-  }
-}
-
-export default User;
