+++ client/src/pages/AgeGate.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const AgeGate = () => {
+  return (
+    <p>This is our age gate</p>
+  )
+}
+
+export default AgeGate;
