+++ client/src/pages/User.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class User extends React.Component {
+  render() {
+    return (
+      <main>
+        <p>This is the user page</p>
+      </main>
+    )
+  }
+}
+
+export default User;
